Software/Python GUI/requirements.txt
numpy>=1.24
pyserial>=3.5
pyqt5>=5.15.7
matplotlib>=3.7
pandas>=1.5
scipy>=1.10